# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_MoranWeissman_sharko_internal_orchestrator.SecretMatch:
    properties:
      field:
        type: string
      line:
        description: |-
          Line is the 1-indexed line number where the match occurred. Helps
          the user open the file and find the field. Zero if unknown.
        type: integer
      pattern:
        type: string
    type: object
...
  internal_api.aiAnnotateBlockedResponse:
    properties:
      code:
        type: string
      matches:
        items:
          $ref: '#/definitions/github_com_MoranWeissman_sharko_internal_orchestrator.SecretMatch'
        type: array
      message:
        type: string
    type: object
...
  internal_api.annotateAddonValuesResponse:
    properties:
      ai_skip_reason:
        description: |-
          AISkipReason is set when the LLM was skipped and the heuristic-only
          output was committed instead. Possible values: "not_configured",
          "opted_out", "oversize", "timeout", "llm_error", "parse_error".
        type: string
      branch:
        type: string
      commit_sha:
        type: string
      merged:
        type: boolean
      pr_id:
        type: integer
      pr_url:
        type: string
    type: object
...
  internal_api.setAIOptOutRequest:
    properties:
      opt_out:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Addon management server for Kubernetes clusters, built on ArgoCD.
  title: Sharko API
  version: 1.0.0
paths:
...
  /addons/{name}/values/ai-opt-out:
    put:
      consumes:
      - application/json
      description: 'Sets or clears the `# sharko: ai-annotate=off` directive in the
        addon''s global values file header. When opt-out is true, future addon-add
        and refresh-from-upstream runs skip the AI annotate pass for this addon. The
        directive is the only persistence (no Sharko-side store). Idempotent.'
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      - description: opt_out flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.setAIOptOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PR opened (or merged) with header-only change
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Addon not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Toggle per-addon AI annotation opt-out
      tags:
      - addons
  /addons/{name}/values/annotate:
    post:
      description: Fetches the chart's upstream values.yaml at the catalog-pinned
        version, runs the V121-7 AI annotate pass (one-line `# description` comments
        + LLM-suggested cluster-specific paths unioned with the heuristic), and opens
        a Tier 2 PR with the result. Hard-blocked when the secret-leak guard matches;
        in that case returns 422 with redacted matches and no PR. Skipped when AI
        is not configured or the addon is opted out.
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PR opened (or merged if auto-merge is on)
          schema:
            $ref: '#/definitions/internal_api.annotateAddonValuesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Addon not found
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Secret-leak guard blocked the LLM call
          schema:
            $ref: '#/definitions/internal_api.aiAnnotateBlockedResponse'
        "502":
          description: Git or upstream fetch error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: AI not configured
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Re-annotate an addon's global values via AI
      tags:
      - addons
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"